+++ packages/components/src/base/Accordion/accordionactions.tsx
@@ -1,5 +1,5 @@
 import { AccordionActions as MuiAccordionActions, type AccordionActionsProps } from '@mui/material';
 
-export function AccordionActions(props: AccordionActionsProps) {
+export function AccordionActions(props: AccordionActionsProps): JSX.Element {
   return <MuiAccordionActions {...props} />;
 }