+++ packages/components/src/base/Avatar/avatar.tsx
@@ -1,5 +1,5 @@
 import { Avatar as MuiAvatar, type AvatarProps } from '@mui/material';
 
-export function Avatar(props: AvatarProps) {
+export function Avatar(props: AvatarProps): JSX.Element {
   return <MuiAvatar {...props} />;
 }