+++ packages/components/src/base/Button/iconbutton.tsx
@@ -1,5 +1,5 @@
 import { IconButton as MuiIconButton, type IconButtonProps } from '@mui/material';
 
-export function IconButton(props: IconButtonProps) {
+export function IconButton(props: IconButtonProps): JSX.Element {
   return <MuiIconButton {...props} />;
 }