+++ packages/components/src/base/Chip/chip.tsx
@@ -1,5 +1,5 @@
 import { Chip as MuiChip, type ChipProps } from '@mui/material';
 
-export function Chip(props: ChipProps) {
+export function Chip(props: ChipProps): JSX.Element {
   return <MuiChip {...props} />;
 }