+++ packages/components/src/base/Dialog/dialogactions.tsx
@@ -1,5 +1,5 @@
 import { DialogActions as MuiDialogActions, type DialogActionsProps } from '@mui/material';
 
-export function DialogActions(props: DialogActionsProps) {
+export function DialogActions(props: DialogActionsProps): JSX.Element {
   return <MuiDialogActions {...props} />;
 }