+++ packages/components/src/base/Form/formcontrollabel.tsx
@@ -1,5 +1,5 @@
 import { FormControlLabel as MuiFormControlLabel, type FormControlLabelProps } from '@mui/material';
 
-export function FormControlLabel(props: FormControlLabelProps) {
+export function FormControlLabel(props: FormControlLabelProps): JSX.Element {
   return <MuiFormControlLabel {...props} />;
 }