+++ packages/components/src/base/Menu/menu.tsx
@@ -1,5 +1,5 @@
 import { Menu as MuiMenu, type MenuProps } from '@mui/material';
 
-export function Menu(props: MenuProps) {
+export function Menu(props: MenuProps): JSX.Element {
   return <MuiMenu {...props} />;
 }