+++ packages/components/src/base/Popper/popper.tsx
@@ -1,5 +1,5 @@
 import { Popper as MuiPopper, type PopperProps } from '@mui/material';
 
-export function Popper(props: PopperProps) {
+export function Popper(props: PopperProps): JSX.Element {
   return <MuiPopper {...props} />;
 }