+++ packages/components/src/custom/Dialog/StyledDialogActions.tsx
@@ -1,12 +1,12 @@
-import { FC, ReactNode } from 'react';
+import React from 'react';
 import { DialogActions } from '../../base/Dialog';
 
 interface DialogActionsProps {
-  children: ReactNode;
+  children: React.ReactNode;
 }
 
-const MesheryDialogActions: FC<DialogActionsProps> = ({ children, ...props }) => {
+function StyledDialogActions({ children, ...props }: DialogActionsProps): JSX.Element {
   return <DialogActions {...props}>{children}</DialogActions>;
-};
+}
 
-export { MesheryDialogActions as StyledDialogActions };
+export default StyledDialogActions;