+++ packages/components/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.1-beta.10](https://github.com/layer5io/sistent/compare/@layer5/sistent-components@0.8.1-beta.9...@layer5/sistent-components@0.8.1-beta.10) (2023-11-04)
+
+**Note:** Version bump only for package @layer5/sistent-components
+
+## [0.8.1-beta.9](https://github.com/layer5io/sistent/compare/@layer5/sistent-components@0.8.1-beta.8...@layer5/sistent-components@0.8.1-beta.9) (2023-11-04)
+
+### Bug Fixes
+
+- redo const to function ([056f675](https://github.com/layer5io/sistent/commit/056f675fdbd4c964ab07f4ce06b266d7dab53112))
+
+## [0.8.1-beta.8](https://github.com/layer5io/sistent/compare/@layer5/sistent-components@0.8.1-beta.7...@layer5/sistent-components@0.8.1-beta.8) (2023-11-03)
+
+### Bug Fixes
+
+- remove react-error-boundary package ([6ea9422](https://github.com/layer5io/sistent/commit/6ea9422a8125b67d1241dd6a1c323f8830a509a3))
+
+## [0.8.1-beta.7](https://github.com/layer5io/sistent/compare/@layer5/sistent-components@0.8.1-beta.6...@layer5/sistent-components@0.8.1-beta.7) (2023-11-03)
+
+### Bug Fixes
+
+- **components:** refactor components ([5b90831](https://github.com/layer5io/sistent/commit/5b90831f02644bbaa16de6e59a11a3ea50f17e58))
+
 ## [0.8.1-beta.6](https://github.com/layer5io/sistent/compare/@layer5/sistent-components@0.8.1-beta.5...@layer5/sistent-components@0.8.1-beta.6) (2023-11-02)
 
 **Note:** Version bump only for package @layer5/sistent-components